+++ flake8_trio.py
@@ -29,7 +29,7 @@ from typing import (
 from flake8.options.manager import OptionManager
 
 # CalVer: YY.month.patch, e.g. first release of July 2022 == "22.7.1"
-__version__ = "22.8.8"
+__version__ = "22.9.1"
 
 
 Error_codes = {